refactor(friends): Moved FB friend get logic into view model.
