+++ app/src/main/java/com/accountabilibuddies/accountabilibuddies/fragments/AddFriendsFragment.java
@@ -5,36 +5,25 @@ import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
 import android.view.LayoutInflater;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 
 import com.accountabilibuddies.accountabilibuddies.R;
 import com.accountabilibuddies.accountabilibuddies.databinding.FragmentAddFriendsBinding;
-import com.accountabilibuddies.accountabilibuddies.model.Friend;
-import com.accountabilibuddies.accountabilibuddies.network.APIClient;
-import com.parse.ParseUser;
-
-import org.apache.commons.collections4.CollectionUtils;
-
-import java.util.ArrayList;
-import java.util.List;
+import com.accountabilibuddies.accountabilibuddies.viewmodel.AddFriendsViewModel;
 
 public class AddFriendsFragment extends Fragment {
 
-
-    private AppCompatActivity listener;
     private FragmentAddFriendsBinding binding;
+    private AddFriendsViewModel viewModel;
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
 
-        if (context instanceof AppCompatActivity) {
-            this.listener = (AppCompatActivity) context;
-        }
+        viewModel = new AddFriendsViewModel(context);
     }
 
     @Nullable
@@ -42,7 +31,7 @@ public class AddFriendsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup parent, @Nullable Bundle savedInstanceState) {
 
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_add_friends, parent, false);
-        binding.executePendingBindings();
+        binding.setAddFriendsViewModel(viewModel);
 
         setUpFriendsView();
 
@@ -51,33 +40,18 @@ public class AddFriendsFragment extends Fragment {
 
     private void setUpFriendsView() {
 
-        APIClient.getClient().getFriendsByUserId(
-            ParseUser.getCurrentUser().getObjectId(),
-            new APIClient.GetFriendsListener() {
-                @Override
-                public void onSuccess(List<Friend> friends) {
+        viewModel.showFriendsView(binding.actvFriends);
 
-                    List<String> friendNames = new ArrayList(CollectionUtils.collect(friends,
-                            (Friend friend) -> {
-                                return friend.getName();
-                            }
-                    ));
+        binding.actvFriends.setOnTouchListener(
+            (View view, MotionEvent event) -> {
 
-                    String[] friendNameArray = friendNames.toArray(new String[0]);
-
-                    ArrayAdapter<String> adapter = new ArrayAdapter<>(
-                            listener.getBaseContext(),
-                            android.R.layout.simple_dropdown_item_1line,
-                            friendNameArray
-                    );
-
-                    binding.actvFriends.setAdapter(adapter);
+                if (event.getAction() == MotionEvent.ACTION_DOWN) {
+                    binding.actvFriends.showDropDown();
+                    binding.actvFriends.requestFocus();
                 }
 
-                @Override
-                public void onFailure(String errorMessage) {
-
-                }
-            });
+                return true;
+            }
+        );
     }
 }
